Blocks: Break out components into separate files
React best practices tend to use a "one component per file" standard. This makes components easier to conceptualize as stand-alone items, easier to test, and easier to document.
