+++ public_html/wp-content/mu-plugins/blocks/source/blocks/speakers/block-content.js
@@ -12,11 +12,11 @@ import { Component } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { ItemTitle, DangerousItemHTMLContent, ItemPermalink, BlockNoContent } from '../../components/block-content';
-import { AvatarImage }                                               from '../../components/image';
-import { PostList }                                                  from '../../components/post-list';
-import { filterEntities }                                            from '../../data';
-import { tokenSplit, arrayTokenReplace }                             from '../../i18n';
+import { ItemTitle, DangerousItemHTMLContent, ItemPermalink } from '../../components/block-content';
+import { AvatarImage }                                        from '../../components/image';
+import { BlockNoContent, PostList }                           from '../../components/post-list';
+import { filterEntities }                                     from '../../data';
+import { tokenSplit, arrayTokenReplace }                      from '../../i18n';
 
 /**
  * Component for the section of each speaker post that displays information about relevant sessions.
